requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
cffi==1.17.1
cryptography==44.0.0
Django==4.2.16
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1
